server/requirements.txt
anyio==3.6.2
appnope==0.1.3
argcomplete==2.0.0
asttokens==2.2.1
attrs==22.2.0
autopep8==1.7.0
backcall==0.2.0
black==23.1a1
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
datamodel-code-generator==0.16.0
dateparser==1.1.5
debugpy==1.6.5
decorator==5.1.1
dnspython==2.2.1
docutils==0.19
email-validator==1.3.1
entrypoints==0.4
executing==1.2.0
fastapi==0.89.0
Flask==2.2.2
genson==1.2.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-resources==5.10.2
inflect==5.6.2
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-spec==0.1.2
jupyter_client==7.4.9
jupyter_core==5.1.3
lazy-object-proxy==1.9.0
lockfile==0.12.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mock==5.0.0
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.6
note==0.5.2
openapi-schema-validator==0.3.4
openapi-spec-validator==0.5.1
orjson==3.8.1
packaging==21.3
parso==0.8.3
pathable==0.4.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
pluggy==1.0.0
prance==0.22.11.4.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pydantic==1.10.4
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
PySnooper==1.1.1
pytest==7.2.1
python-daemon==2.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
semver==2.13.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.22.0
toml==0.10.2
tornado==6.2
traitlets==5.8.1
typed-ast==1.5.4
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.20.0
vcrpy==4.2.1
watchfiles==0.18.0
wcwidth==0.2.6
websockets==10.3
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
